worker/requirements.txt
celery==5.4.0
redis==5.2.0
pytest-celery==1.1.3
pydantic==2.9.2
pydantic_settings==2.6.1
sqlmodel==0.0.22
numpy==1.26.4
pandas-ta==0.3.14b0
pandas==2.2.3
apimoex==1.4.0
yahoo-fin==0.8.9.1
httpx==0.27.2